# Documentacion/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Vida Cotidiana API
  version: 1.0.0
  description: >
    Canonical API contract for Vida Cotidiana V1. All timestamps are
    ISO-8601 in UTC. All identifiers are UUIDv4. Authentication is delegated
    to Keycloak (OIDC/OAuth 2.1, Authorization Code + PKCE, DEC-004/ADR-008);
    this API is a pure OAuth2 resource server and defines no /auth/* login
    endpoint of its own. An operational healthcheck (e.g. Spring Boot
    Actuator /actuator/health) exists outside this versioned business
    contract and is intentionally not documented here (see
    07-backend-architecture.md).
servers:
  - url: /api/v1
security:
  - keycloakBearer: []
paths:
...
  /maintenance-records:
    get:
      summary: List own maintenance records (paginated) — BE-037/WEB-009, V2 (see 22-decision-log.md ADR-015 refinamiento 2026-08-18).
      security:
        - keycloakBearer: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated maintenance record list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    additionalProperties: true # BE-035, see PageMeta
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/MaintenanceRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a maintenance record (caller becomes owner)
      security:
        - keycloakBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaintenanceRecordRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /maintenance-records/{id}:
    get:
      summary: Get a maintenance record (owner only — no sharing/collaborator concept for this entity)
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: MaintenanceRecord
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update maintenance record (owner only). Requires the current version for optimistic locking.
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMaintenanceRecordRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict (optimistic locking) — the maintenance record was modified concurrently; refetch and retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete maintenance record (owner only)
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /maintenance-records/{id}/complete:
    post:
      summary: Toggle the completion state (owner only). Optionally accepts the expected version for optimistic locking.
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteMaintenanceRecordRequest'
      responses:
        '200':
          description: Completion state toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict (optimistic locking)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /commitments:
    get:
      summary: List own commitments (paginated), optionally filtered by direction — ADR-016/FR-025/FR-027, V3. Módulo Laboral.
      security:
        - keycloakBearer: []
      parameters:
        - in: query
          name: direction
          required: false
          schema:
            type: string
            enum: [MINE, THEIRS]
          description: >
            "Seguimientos" (MINE) / "Esperando" (THEIRS) — misma entidad,
            filtrada. Omitido devuelve ambas direcciones juntas.
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated commitment list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    additionalProperties: true
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Commitment'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a commitment (caller becomes owner)
      security:
        - keycloakBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommitmentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commitment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: personId/projectId/originReminderId does not reference a resource owned by the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /commitments/{id}:
    get:
      summary: Get a commitment (owner only)
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Commitment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commitment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update commitment (owner only). Requires the current version for optimistic locking.
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommitmentRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commitment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict (optimistic locking)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete commitment (owner only)
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /commitments/{id}/resolve:
    post:
      summary: Mark a commitment resolved (owner only, UC-20). Optionally accepts the expected version for optimistic locking.
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveCommitmentRequest'
      responses:
        '200':
          description: Commitment resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commitment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict (optimistic locking)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    keycloakBearer:
      type: oauth2
      description: >
        OAuth2/OIDC access token issued by Keycloak (Authorization Code +
        PKCE from the client). The backend validates the bearer JWT on every
        request as a pure OAuth2 resource server; it never issues its own
        session or token (DEC-004/ADR-008). Token issuer/JWK Set URL are
        environment configuration, not part of this contract.
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.example/realms/vida-cotidiana/protocol/openid-connect/auth
          tokenUrl: https://keycloak.example/realms/vida-cotidiana/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect authentication
  parameters:
    PageParam:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page index.
    SizeParam:
      in: query
      name: size
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Page size (max 100).
  schemas:
    Error:
      type: object
      description: Uniform error envelope. Never includes stack traces, SQL, internal names or secrets.
      required: [code, message, traceId]
      properties:
        code:
          type: string
          example: REMINDER_NOT_FOUND
        message:
          type: string
          example: The requested reminder was not found.
        traceId:
          type: string
    PageMeta:
      type: object
      # BE-035: additionalProperties: true is required here because PageMeta is combined via allOf
      # with a sibling object (page items) at each paginated endpoint — without it, strict allOf
      # validators (incl. swagger-request-validator/TEST-API-001) reject each branch as a closed
      # schema and fail the sibling's fields as "not allowed", even though the merged instance is
      # exactly what the composed schema intends. Not a loosening of type/required validation.
      additionalProperties: true
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...
    CreateMaintenanceRecordRequest:
      type: object
      required: [item, nextDueAt]
      properties:
        item:
          type: string
          minLength: 1
          maxLength: 200
        nextDueAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp in UTC.
    UpdateMaintenanceRecordRequest:
      type: object
      description: Partial update. version is required to detect concurrent modifications (optimistic locking).
      required: [version]
      properties:
        item:
          type: string
          minLength: 1
          maxLength: 200
        nextDueAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp in UTC.
        version:
          type: integer
          description: Expected current version of the maintenance record; a mismatch returns 409.
    CompleteMaintenanceRecordRequest:
      type: object
      description: Optional. If version is provided, the server validates it for optimistic locking; if omitted, the toggle is applied without a concurrency check.
      properties:
        version:
          type: integer
    MaintenanceRecord:
      type: object
      required: [id, item, nextDueAt, status, version]
      description: >
        BE-037/WEB-009 (V2, see 22-decision-log.md ADR-015 refinamiento
        2026-08-18). Owner-only entity — no sharing/collaborator concept.
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        item:
          type: string
          minLength: 1
          maxLength: 200
        nextDueAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp in UTC.
        status:
          type: string
          enum: [AL_DIA, PROXIMO, VENCIDO, COMPLETADO]
          description: >
            Derived at response time from nextDueAt vs "now" (ASSUMPTION:
            PROXIMO = within 30 days of nextDueAt) unless the record was
            explicitly marked completed, in which case it is always
            COMPLETADO regardless of nextDueAt. Never a stored temporal
            value — see Documentacion/09-data-model.md.
        version:
          type: integer
          description: Optimistic locking version, incremented on every update.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateCommitmentRequest:
      type: object
      required: [personId, description, direction, dueAt]
      description: >
        ADR-016/FR-025. "Seguimientos" (direction=MINE) y "Esperando"
        (direction=THEIRS) son la misma entidad, distinguida por dirección —
        no dos recursos separados.
      properties:
        personId:
          type: string
          format: uuid
          description: >
            Obligatorio (ASSUMPTION, no decidida por el Product Owner — ver
            ADR-016 TBD explícitos). Debe pertenecer al mismo dueño.
        description:
          type: string
          minLength: 1
          maxLength: 2000
        direction:
          type: string
          enum: [MINE, THEIRS]
        dueAt:
          type: string
          format: date-time
        projectId:
          type: string
          format: uuid
          description: Opcional — debe pertenecer al mismo dueño; si no, 404 PROJECT_NOT_FOUND.
        originReminderId:
          type: string
          format: uuid
          description: Opcional — el Reminder/reunión del que surgió (UC-21); debe pertenecer al mismo dueño.
    UpdateCommitmentRequest:
      type: object
      description: >
        Partial update. version is required to detect concurrent
        modifications (optimistic locking). Cubre UC-20 "reprogramar"
        (enviar solo dueAt) y el cambio de dirección (ADR-016: el sentido
        completo del modelo unificado, no una acción especial).
      required: [version]
      properties:
        personId:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 2000
        direction:
          type: string
          enum: [MINE, THEIRS]
        dueAt:
          type: string
          format: date-time
        projectId:
          type: string
          format: uuid
        version:
          type: integer
          description: Expected current version of the commitment; a mismatch returns 409.
    ResolveCommitmentRequest:
      type: object
      description: Optional. If version is provided, the server validates it for optimistic locking; if omitted, resolve is applied without a concurrency check.
      properties:
        version:
          type: integer
    Commitment:
      type: object
      required: [id, personId, description, direction, dueAt, status, version]
      description: ADR-016/FR-025/FR-027. Owner-only entity — no sharing/collaborator concept.
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        personId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        description:
          type: string
          minLength: 1
          maxLength: 2000
        direction:
          type: string
          enum: [MINE, THEIRS]
        dueAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [OPEN, DONE]
        originReminderId:
          type: string
          format: uuid
        version:
          type: integer
          description: Optimistic locking version, incremented on every update.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Input validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Not authenticated (missing or invalid bearer token).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource does not exist or is not accessible to the caller (never distinguishes the two, to avoid leaking existence, SEC-001).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...